Simplify Calypso
The read contract was never being called, it was only ever being called
from inside the write contract. With this change, we keep the read contract
as a way to differentiate instances that are holding the Read and the Write
structures, but we do away with the contract code and it's registration.

Also, fix documentation formatting.